// goody/Week3/mission3/src/types/movie.ts
export type Movie = {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
};

export type MovieResponse = {
    page: number;
    results : Movie[];
    total_pages : number;
    total_result: number;
};

export type MovieDetails = {
    adult: boolean;
    backdrop_path: string;
    budget: number;
    genres: [];
    homepage: string;
    id: number; 
    imdb_id: string;
    origin_country: [];
    original_language: string;
    original_title: string;
    overview: string;
    poster_path: string;
    production_companies: [];
    production_countries: [];
    release_date: string;
    revenue: number;
    runtime: number;
    spoken_languages: [];
    status: string;
    tagline: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
    }
    
export type Credit = {
    adult : boolean;
    gender : number;
    id : number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string;
    cast_id: number;
    credit_id: string;

    order? : number;
    character?: string;

    department?: string;
    job?: string;
}

export type CreditResponse = { 
    cast: Credit[];
    crew: Credit[];
    id : number;
}